requirements.txt
python-fasthtml==0.6.9
httpx==0.27.2
tree-sitter==0.20.1